backend/requirements.txt
Django==4.1
djangorestframework==3.14.0
django-cors-headers==3.13.0
drf-spectacular==0.24.1
black==22.12.0
selfies==0.2.0

-e .

-e ./science
 